requirements.txt
Flask==2.1.0
gunicorn==20.1.0
requests~=2.31.0
Pillow~=10.0.0
market~=0.0